requirements.txt
absl-py==0.11.0
astroid==2.4.2
astunparse==1.6.3
attrs==20.3.0
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
graphviz==0.5.1
grpcio==1.32.0
h5py==2.10.0
idna==2.10
isort==5.6.4
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
matplotlib==3.3.3
mccabe==0.6.1
neupy==0.6.5
numpy==1.19.3
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.0.1
progressbar2==3.34.3
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pyparsing==2.4.7
pyswarms==1.2.0
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scipy==1.5.4
six==1.15.0
tableprint==0.7.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
Theano==1.0.0
toml==0.10.2
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
